version 0.2.3 (eff bundler)
Let me tell you a story. There's some gem on rubygems that already uses the name 'fml'.
Bundler is supposed to use the order of your sources to determine which gem to use,
and I had set them properly so that bundler preferred our fml to the one on rubygems.

However, now it has apparently, and for no reason AFAICT, decided to prefer rubygems'
fml. I'm now going to try updating to 0.2.3 to see if it will figure itself back out,
even though I really hate this plan.